Added GPS_Connector device
Added GPS_Connector device to the SparkFun-connectors library. This is
used to replace the EM-406 device used previoulsy. This is a more
genaric version of the vertical 6-pin JST connector used on many GPS
modules. This part was used specifically on the GPS Breakout board to
allow use with more modules than just the EM-406.